+++ WinCalcPro/WinCalcPro/Form2.resx
